distsqlrun: expose MockDistSQLServer for testing in other packages
Release note: None
